requirements.txt
requests[tls]>=2.28.0
python-dateutil>=2.8.2
tqdm>=4.65.0
